requirements.txt
requests
pycryptodome==3.10.1
urllib3
tqdm
aiohttp